replace non needed handler
