--- src/modules/transaction/components/payloadDetails/SdaNewHandler.vue
@@ -1,168 +1,0 @@
-<template>
-  <div :class="styleClass">
-    <div>
-      <div>{{ label }}</div>
-      <div>
-        
-        <router-link
-          :to="{
-            name: 'ViewAsset',
-            params: { id: displayValue },
-          }"
-          class="hover:text-blue-primary hover:underline text-blue-600"
-          >{{ displayValue }} 
-        </router-link>
-
-        <span v-if="alias"> 
-          (alias of 
-          <router-link
-          :to="{
-            name: 'ViewNamespace',
-            params: { namespaceParam: alias },
-          }"
-          class="hover:text-blue-primary hover:underline text-blue-600"
-          >{{ alias }}
-          </router-link>
-          )
-        </span>
-        <div v-if="isResolving" class="inline-block animate-spin rounded-full h-3 w-3 border-b-2 border-navy-primary mr-2"></div>
-        <span class="text-sm text-red-300 px-2" v-if="errorMsg">{{ errorMsg }}</span>
-      </div>
-    </div>
-  </div>
-</template>
-
-<script setup lang="ts">
-import { ref, computed, watch } from "vue";
-import { AppState } from "@/state/appState";
-import { UInt64, MosaicId } from "tsjs-xpx-chain-sdk";
-
-const displayValue = ref("");
-const errorMsg = ref("");
-const alias = ref("");
-const isResolving = ref(false);
-
-const props = defineProps({
-  label: String,
-  value: String,
-  styleClass: String,
-  toggleResolve: Boolean,
-  secondaryValue: {
-    type: String,
-    required: true
-  }
-});
-
-const initAssign = ()=>{
-  displayValue.value = props.value || "";
-}
-
-initAssign();
-
-const resolve = async ()=>{
-
-  if(displayValue.value !== ""){
-
-    isResolving.value = true;
-    errorMsg.value = "";
-
-    try {
-      let assetId = new MosaicId(displayValue.value);
-
-      let assetInfo = await AppState.chainAPI!.assetAPI.getMosaic(assetId);
-
-      let tokenDecimalPlace = assetInfo.divisibility;
-      let divider = Math.pow(10, tokenDecimalPlace);
-      let quotient = BigInt(props.secondaryValue)/ BigInt(divider);
-      let remainder = displayValue.value.slice(-tokenDecimalPlace);
-      
-      let remainderInteger = parseInt(remainder);
-
-      let amountWithDecimal = remainderInteger ? remainderInteger / divider: 0;
-      let adjustedDecimalAmount = remainderInteger ? amountWithDecimal.toString().split(".")[1].toString(): "";
-
-      displayValue.value = quotient.toString() + (remainderInteger ? "." + adjustedDecimalAmount :"");
-
-      let names = await AppState.chainAPI!.assetAPI.getMosaicsNames([assetId]);
-      
-      displayValue.value = displayValue.value + " " + props.value;
-
-      if(names[0] && names[0].names[0]){
-        alias.value = names[0].names[0].name;
-      }
-      else{
-        alias.value = "";
-      }
-      
-    } catch (error) {
-      errorMsg.value = "unable to resolve";
-    } 
-  }
-  isResolving.value = false;
-}
-
-if(props.toggleResolve){
-  resolve();
-}
-
-watch(
-  () => props.toggleResolve,
-  (isToggleResolve) => {
-    if(isToggleResolve){
-      resolve();
-    }
-    else{
-      initAssign(); 
-    }
-  }
-)
-
-</script>
-
-<!-- Add "scoped" attribute to limit CSS to this component only -->
-<style scoped lang="scss">
-.txn-div,
-.details {
-  @apply text-gray-800 text-xs;
-
-  > div {
-    @apply flex items-center border-b border-gray-100 py-4;
-
-    > div:first-child {
-      @apply w-40 text-xs pl-4;
-    }
-
-    > div:nth-child(2) {
-      @apply text-xs;
-    }
-
-    > div:last-child {
-      @apply border-none;
-    }
-  }
-}
-
-.table_div {
-  @apply text-xs;
-
-  > div {
-    @apply grid grid-cols-4;
-
-    > div {
-      @apply p-2;
-    }
-
-    > div:first-child {
-      @apply text-blue-primary font-bold;
-    }
-
-    > div:nth-child(2) {
-      @apply break-all col-span-3;
-    }
-  }
-
-  > div:nth-child(2n + 1) {
-    @apply bg-gray-100;
-  }
-}
-</style>
